// types/models.ts
export interface TrainerRowData {
    trainerName: string;
    username: string;
    trainerCode: string;
    altTrainerName?: string;
    altTrainerCode?: string;
    altTrainerName2?: string;
    altTrainerCode2?: string;
    altTrainerName3?: string;
    altTrainerCode3?: string;
    altTrainerName4?: string;
    altTrainerCode4?: string;
}